rouge: allow directories to be used to fill partition
This commit allows to specify directories in `items`
for ext4 and vfat partitions.
For example:

```
mypart:
    gpt_type: EBD0A0A2-B9E5-4433-87C0-68B6B72699C7
    type: ext4
    #type: vfat
    folders:
        "/": "/mnt/builds/tmp/x"
        "aa": "/mnt/builds/tmp/x"
        "/cc/dd/": "/mnt/builds/tmp/x"
```

Symlinks are supported inside directories for ext4 partitions,
and are copied as is.
For vfat symlinks are skipped, and warning is printed.

Acked-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>

+++ moulin/rouge/block_entry.py
@@ -246,10 +246,24 @@ class FileSystem(BlockEntry):
 
     def _complete_init(self):
         for _, local_name, local_mark in self._items:
-            if not os.path.isfile(local_name):
+            if not os.path.isfile(local_name) and not os.path.isdir(local_name):
                 raise YAMLProcessingError(f"Can't find file '{local_name}'", local_mark)
 
-        files_size = sum([os.path.getsize(x[1]) for x in self._items]) + 8 * 1024 * 1024
+        # calculate size of original files and directories
+        files_size = 0
+        for remote_path, local_path, _ in self._items:
+            if os.path.isfile(local_path):
+                files_size += os.path.getsize(local_path)
+            if os.path.isdir(local_path):
+                files_size += os.path.getsize(local_path)
+                for (dirpath, dirnames, filenames) in os.walk(local_path, topdown=True):
+                    for filename in filenames:
+                        # we may have links to location that is incorrect on host, so we use lstat to handle this
+                        files_size += os.lstat(os.path.join(dirpath, filename)).st_size
+                    for dirname in dirnames:
+                        files_size += os.path.getsize(os.path.join(dirpath, dirname))
+        files_size += 8 * 1024 * 1024
+
         size_node = self._node.get("size", None)
         if size_node:
             self._size = _parse_size(size_node)
@@ -291,7 +305,10 @@ class Ext4(FileSystem):
                     # create destination subfolder
                     os.makedirs(os.path.join(tempd, remote_path_and_name[0]), exist_ok=True)
 
-                shutil.copyfile(local, os.path.join(tempd, remote))
+                if os.path.isfile(local):
+                    shutil.copyfile(local, os.path.join(tempd, remote))
+                if os.path.isdir(local):
+                    shutil.copytree(local, os.path.join(tempd, remote), symlinks=True, dirs_exist_ok=True)
             tempf.truncate(self._size)
             ext_utils.mkext4fs(tempf, tempd)
             ext_utils.dd(tempf, fp, offset)
@@ -299,12 +316,34 @@ class Ext4(FileSystem):
 
 class Vfat(FileSystem):
     "Represents vfat fs with list of files"
+
+    def unwrap_dirs(self):
+        "Return list of files with flatten content of the directories"
+        out_list: List[Tuple[str, str, Mark]] = []
+        for remote, local, mark in self._items:
+            if os.path.isfile(local):
+                out_list.append([remote, local, mark])
+            if os.path.isdir(local):
+                for (dirpath, _, filenames) in os.walk(local, topdown=True):
+                    remote_dirpath = dirpath.replace(local, remote, 1)
+                    for filename in filenames:
+                        # we skip symlinks as not supported on vfat
+                        if os.path.islink(os.path.join(dirpath, filename)):
+                            log.warn("Symlink '%s' is skipped.", os.path.join(dirpath, filename))
+                        else:
+                            out_list.append([os.path.join(remote_dirpath, filename),
+                                             os.path.join(dirpath, filename), mark])
+        return out_list
+
     def write(self, fp, offset):
         if not self._size:
             self._complete_init()
         with NamedTemporaryFile() as tempf:
             tempf.truncate(self._size)
             ext_utils.mkvfatfs(tempf)
+            # for vfat we have to create each subdir (if any) and copy file one by one
+            # that's why we need to 'unwrap' content of any input directory
+            self._items = self.unwrap_dirs()
             # scan all remote filenames and collect the list of folders to create
             list_for_mmd = list()
             for remote, _, _ in self._items:
